# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GitHub Test Case Generator API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api/v1" }
  ],
  "paths": {
...
    "/testcases/execute": {
      "post": {
        "summary": "Execute tests in sandbox and return structured results",
        "requestBody": { "required": true },
        "responses": { "200": { "description": "Execution results" } }
      }
    },
...
  },
...
}